+++ mobile/BubbleWar.Android/packages.config
@@ -7,6 +7,8 @@
   <package id="Microsoft.AppCenter.Crashes" version="1.12.0" targetFramework="monoandroid90" />
   <package id="Newtonsoft.Json" version="12.0.1" targetFramework="monoandroid90" />
   <package id="Plugin.CurrentActivity" version="2.1.0.4" targetFramework="monoandroid90" />
+  <package id="Polly" version="6.1.2" targetFramework="monoandroid90" />
+  <package id="Refit" version="4.6.58" targetFramework="monoandroid90" />
   <package id="Syncfusion.Licensing" version="16.4.0.46" targetFramework="monoandroid90" />
   <package id="Syncfusion.Xamarin.Core" version="16.4.0.46" targetFramework="monoandroid90" />
   <package id="Syncfusion.Xamarin.SfChart" version="16.4.0.46" targetFramework="monoandroid90" />
