+++ mobile/BubbleWar/Models/MutationData.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace BubbleWar
+{
+    class MutationData
+    {
+        [JsonProperty("incrementPoints")]
+        public TeamScore TeamScore { get; set; }
+    }
+}
