+++ mobile/BubbleWar/Models/MutationResponse.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace BubbleWar
+{
+    class MutationResponse : ErrorResponse
+    {
+        [JsonProperty("data")]
+        public MutationData Data { get; set; }
+    }
+}
