+++ mobile/BubbleWar/Models/QueryRequest.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace BubbleWar
+{
+    class QueryRequest
+    {
+        [JsonProperty("query")]
+        public string Query => "{teams{id, name, points}}";
+    }
+}
