+++ mobile/BubbleWar/Models/QueryResponse.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace BubbleWar
+{
+    class QueryResponse : ErrorResponse
+    {
+        [JsonProperty("data")]
+        public ResponseData Data { get; set; }
+    }
+}
