+++ mobile/BubbleWar/Models/ResponseData.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections.Generic;
+
+using Newtonsoft.Json;
+
+namespace BubbleWar
+{
+    class ResponseData
+    {
+        [JsonProperty("teams")]
+        public List<TeamScore> Teams { get; set; }
+    }
+}
