+++ mobile/BubbleWar/Services/GraphQLService.cs
@@ -1,15 +1,19 @@
-﻿using System.Collections.Generic;
-using System.Net.Http;
+﻿using System;
+using System.Linq;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
-using Newtonsoft.Json;
-using System.Linq;
-using System;
+using Refit;
+using Polly;
 
 namespace BubbleWar
 {
-    public abstract class GraphQLService : BaseGraphQLService
+    public abstract class GraphQLService
     {
+        #region Constant Fields
+        readonly static IGraphQLAPI _graphQLApiClient = RestService.For<IGraphQLAPI>(BubbleWarUrl);
+        #endregion
+
         #region Properties
         static string BubbleWarUrl => GraphQLSettings.Uri.ToString();
         #endregion
@@ -17,7 +21,7 @@ namespace BubbleWar
         #region Methods
         public static async Task<List<TeamScore>> GetTeamScoreList()
         {
-            var response = await PostObjectToAPI<QueryRequest, QueryResponse>(BubbleWarUrl, new QueryRequest()).ConfigureAwait(false);
+            var response = await ExecutePollyFunction(() => _graphQLApiClient.Query(new QueryRequest())).ConfigureAwait(false);
 
             if (response.Errors != null)
                 throw new AggregateException(response.Errors.Select(x => new Exception(x.Message)));
@@ -27,7 +31,7 @@ namespace BubbleWar
 
         public static async Task<TeamScore> VoteForTeamAndGetCurrentScore(TeamColor teamType)
         {
-            var response = await PostObjectToAPI<MutationRequest, MutationResponse>(BubbleWarUrl, new MutationRequest(teamType)).ConfigureAwait(false);
+            var response = await ExecutePollyFunction(() => _graphQLApiClient.Mutation(new MutationRequest(teamType))).ConfigureAwait(false);
 
             if (response.Errors != null)
                 throw new AggregateException(response.Errors.Select(x => new Exception(x.Message)));
@@ -36,71 +40,18 @@ namespace BubbleWar
         }
 
         public static Task VoteForTeam(TeamColor teamType) => VoteForTeamAndGetCurrentScore(teamType);
-        #endregion
 
-        #region Classes
-        class QueryRequest
+        static Task<T> ExecutePollyFunction<T>(Func<Task<T>> action, int numRetries = 3)
         {
-            [JsonProperty("query")]
-            public string Query => "{teams{id, name, points}}";
-        }
+            return Policy
+                    .Handle<Exception>()
+                    .WaitAndRetryAsync
+                    (
+                        numRetries,
+                        pollyRetryAttempt
+                    ).ExecuteAsync(action);
 
-        class QueryResponse : ErrorResponse
-        {
-            [JsonProperty("data")]
-            public ResponseData Data { get; set; }
-        }
-
-        class ResponseData
-        {
-            [JsonProperty("teams")]
-            public List<TeamScore> Teams { get; set; }
-        }
-
-        class MutationRequest
-        {
-            readonly int _teamNumber;
-
-            public MutationRequest(TeamColor team) => _teamNumber = (int)team;
-
-            [JsonProperty("query")]
-            public string Query => "mutation{incrementPoints(id:" + _teamNumber + ") {id, name, points}}";
-        }
-
-        class MutationResponse : ErrorResponse
-        {
-            [JsonProperty("data")]
-            public MutationData Data { get; set; }
-        }
-
-        class MutationData
-        {
-            [JsonProperty("incrementPoints")]
-            public TeamScore TeamScore { get; set; }
-        }
-
-        abstract class ErrorResponse
-        {
-            [JsonProperty("errors")]
-            public List<Error> Errors { get; set; }
-        }
-
-        class Error
-        {
-            [JsonProperty("message")]
-            public string Message { get; }
-
-            [JsonProperty("locations")]
-            public List<ErrorLocations> Locations { get; set; }
-        }
-
-        class ErrorLocations
-        {
-            [JsonProperty("line")]
-            public long Line { get; set; }
-
-            [JsonProperty("column")]
-            public long Column { get; set; }
+            TimeSpan pollyRetryAttempt(int attemptNumber) => TimeSpan.FromSeconds(Math.Pow(2, attemptNumber));
         }
         #endregion
     }
