+++ mobile/BubbleWar/Services/IGraphQLAPI.cs
@@ -1,0 +1,16 @@
+﻿using System.Threading.Tasks;
+using Refit;
+
+namespace BubbleWar
+{
+    [Headers("Accept-Encoding: gzip",
+                "Accept: application/json")]
+    interface IGraphQLAPI
+    {
+        [Post("")]
+        Task<QueryResponse> Query([Body] QueryRequest request);
+
+        [Post("")]
+        Task<MutationResponse> Mutation([Body] MutationRequest request);
+    }
+}
