+++ mobile/BubbleWar/Services/MutationRequest.cs
@@ -1,0 +1,14 @@
+﻿using Newtonsoft.Json;
+
+namespace BubbleWar
+{
+    class MutationRequest
+    {
+        readonly int _teamNumber;
+
+        public MutationRequest(TeamColor team) => _teamNumber = (int)team;
+
+        [JsonProperty("query")]
+        public string Query => "mutation{incrementPoints(id:" + _teamNumber + ") {id, name, points}}";
+    }
+}
